+++ attributes/php_driver.rb
@@ -8,10 +8,10 @@
 #
 
 # latest build numbers for each CUBRID PHP version in the form of 'version'=>'build_number'
-build_numbers = {'9.0.0' => '0001', '8.4.3' => '0001', '8.4.1' => '0006'}
+build_numbers = {'9.1.0' => '0001', '9.0.0' => '0001', '8.4.3' => '0001', '8.4.1' => '0006'}
 
 # the default version of CUBRID to install
-default['cubrid']['version'] = "9.0.0"
+default['cubrid']['version'] = "9.1.0"
 
 # the version of a CUBRID PHP driver to install from PECL
 set['cubrid']['php_version'] = "#{node['cubrid']['version']}.#{build_numbers[node['cubrid']['version']]}"